Link Favorites component with store using RTK hooks. WIP
